Add files via upload
Being able to create a script that allows you to recognize a task and the process while prioritizing it based off level of severity would be pivotal in handling issues related to networking outages or possible security breaches. It can be expanded on prioritization of the incidents and how many retries you want to allow the user before it shuts them out to not over consume resources.